completed refactor
