+++ controller/reportCtrl.js
@@ -1,0 +1,53 @@
+import * as RecordCoopDB from '../db/record-coopDB.js';
+import * as CoopDB from '../db/coopDB.js';
+import * as RecordHatchDB from '../db/record-hatchDB.js';
+import * as RecordBroderDB from '../db/record-brooderDB.js';
+import * as EggsDB from '../db/eggsDB.js';
+
+// Get Data for Report
+export const getReportPage = async (req, res) => {
+  const monthlyEggs = await RecordCoopDB.getNumEggsMonthly();
+  const numOfChicken = await CoopDB.getNumberOfChicken();
+  const coopData = await CoopDB.getAllChicken();
+  const eggData = await RecordCoopDB.getCurrMonthEggs();
+  const asOfEggData = await EggsDB.getAsOfTotalEggs();
+  const firstDateCoopRecord = await RecordCoopDB.getFirstDateCoopRecord();
+  const avgEggDaily = await RecordCoopDB.avgDailyEgg();
+  const chickenDeadData = await RecordCoopDB.getChickenDeadCurrMonth();
+  const dailyEggsAMonth = await RecordCoopDB.getDailyEggsInAMonth();
+  const monthlyChickenDead = await RecordCoopDB.getMonthlyChickenDead();
+  const incubationRecord = await RecordHatchDB.getIncubationData();
+  const firstIncubationDate = await RecordHatchDB.getFirstIncubationDate();
+  const totalChickenDeadData = await RecordCoopDB.getTotalChickenDead();
+  const totalIncubationData = await RecordHatchDB.getTotalIncubationData();
+  const dailyChickDeathInAMonth = await RecordBroderDB.getDailyChickDeathInAMonth();
+  const totalChickDeathCurrMonth = await RecordBroderDB.getTotalChickDeathCurrMonth();
+  const cumTotalChickDeath = await RecordBroderDB.getCumTotalChickDeath();
+  const dailyEggsAMonthData = dailyEggsAMonth.map(data => data.numEggs);
+  const monthlyEggsData = monthlyEggs.map((data) => data.numEggs);
+  const monthlyHensDeadData = monthlyChickenDead.map((data) => data.numDeadHen);
+  const monthlyRoosterDeadData = monthlyChickenDead.map((data) => data.numDeadRooster);
+  const dailyChickDeathInAMonthData = dailyChickDeathInAMonth.map((data) => data.numDeadChick);
+  res.status(200)
+    .render('report', {
+      monthlyEggs,
+      numOfChicken,
+      coopData,
+      eggData,
+      asOfEggData,
+      firstDateCoopRecord,
+      avgEggDaily,
+      chickenDeadData,
+      monthlyEggsData,
+      dailyEggsAMonthData,
+      monthlyHensDeadData,
+      monthlyRoosterDeadData,
+      incubationRecord,
+      firstIncubationDate,
+      totalChickenDeadData,
+      totalIncubationData,
+      dailyChickDeathInAMonthData,
+      totalChickDeathCurrMonth,
+      cumTotalChickDeath
+    });
+};
