+++ db/brooderDB.js
@@ -44,23 +44,6 @@ export async function getAllBrooder () {
   }
 }
 
-// Submit Brooder Record
-export async function submitBrooderRecord (brooderData) {
-  try {
-    const brooderID = brooderData.brooderID;
-    const numDeadChick = +brooderData.numDeadChick;
-
-    const result = await pool.query(`
-    INSERT INTO \`record-brooder\` (brooderID, numDeadChick)
-    VALUES (?, ?)
-    `, [brooderID, numDeadChick]);
-    return result;
-  } catch (error) {
-    console.error(error);
-    throw new Error('Error submitting brooder record to database');
-  }
-}
-
 // Update broder mortality rate
 export async function updateBrooderMR (brooderData) {
   try {
