+++ pages/profile.html
@@ -469,7 +469,7 @@
                   <div class="card card-blog card-plain">
                     <div class="card-header p-0 mt-n4 mx-3">
                       <a class="d-block shadow-xl border-radius-xl">
-                        <img src="../assets/img/home-decor-1.jpg" alt="img-blur-shadow" class="img-fluid shadow border-radius-xl">
+                        <img src="../assets/img/dashboard/view-decor-1.jpg" alt="img-blur-shadow" class="img-fluid shadow border-radius-xl">
                       </a>
                     </div>
                     <div class="card-body p-3">
@@ -506,7 +506,7 @@
                   <div class="card card-blog card-plain">
                     <div class="card-header p-0 mt-n4 mx-3">
                       <a class="d-block shadow-xl border-radius-xl">
-                        <img src="../assets/img/home-decor-2.jpg" alt="img-blur-shadow" class="img-fluid shadow border-radius-lg">
+                        <img src="../assets/img/dashboard/view-decor-2.jpg" alt="img-blur-shadow" class="img-fluid shadow border-radius-lg">
                       </a>
                     </div>
                     <div class="card-body p-3">
@@ -543,7 +543,7 @@
                   <div class="card card-blog card-plain">
                     <div class="card-header p-0 mt-n4 mx-3">
                       <a class="d-block shadow-xl border-radius-xl">
-                        <img src="../assets/img/home-decor-3.jpg" alt="img-blur-shadow" class="img-fluid shadow border-radius-xl">
+                        <img src="../assets/img/dashboard/view-decor-3.jpg" alt="img-blur-shadow" class="img-fluid shadow border-radius-xl">
                       </a>
                     </div>
                     <div class="card-body p-3">
