+++ routes/dashboardRoutes.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import * as DashboardCtrl from '../controller/dashboardCtrl.js';
+
+const router = express.Router();
+
+router.get('/view', DashboardCtrl.getDashboardPage);
+
+export default router;
